Add FOL Strat
